Added deleting posts feature using authentication and authorization

+++ controllers/api/v1/posts_api.js
@@ -1,5 +1,6 @@
 const Post = require('../../../models/post');
 const Comment = require('../../../models/comment');
+const { post } = require('../../../routes');
 
 module.exports.index = async function(req, res){
 
@@ -21,16 +22,24 @@ module.exports.index = async function(req, res){
 
 module.exports.destroy = async function(req, res){
     try {
+
         let post = await Post.findById(req.params.id);
+        
+        if(post.user == req.user.id){
 
-        post.remove();
+            post.remove();
 
-        await Comment.deleteMany({post: req.params.id});
+            await Comment.deleteMany({post: req.params.id});
 
 
-        return res.json(200, {
-            message: "Post is deleted Successfully!"
-        });
+            return res.json(200, {
+                message: "Post is deleted Successfully!"
+            });
+        }else{
+            return res.json(401, {
+                message: "You cannot delete this post, Unauthorized User!"
+            });
+        }
 
     } catch (error) {
         return res.json(500, {
